Sun, Jan 23, 2022, 3:04 AM +00:00

+++ CS50x/week1/pset1/credit/tester.c
@@ -5,7 +5,7 @@
 int getLengthOfLong();
 
 int main(void) {
-	long test = get_long("try me, user: ");
+	double test = get_long("try me, user: ");
 	int len = getLengthOfLong(test);
     // long len = test%10;
 	printf("%i\n", len);
